fix: Prevent deletion of null expenses in LeadApiController

+++ src/main/java/site/easy/to/build/crm/controller/api/LeadApiController.java
@@ -71,7 +71,9 @@ public class LeadApiController {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
 
-        expenseServiceImpl.delete(lead.getExpense());
+        if (lead.getExpense() != null ){
+            expenseServiceImpl.delete(lead.getExpense());
+        }
         leadService.delete(lead);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
